upload-monologue/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.12.0
Flask==2.3.3
Werkzeug==3.0.4
